requirements.txt.txt
streamlit
pandas
numpy
scikit-learn
request
pickle
pillow